requirements.txt
# Core
pandas>=2.0.0
numpy>=1.24.0

# Configuration
pyyaml>=6.0

# Statistics
scipy>=1.10.0

# Visualization
matplotlib>=3.8.0

# Development
pytest>=7.3.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
